# public/openapi.yaml
openapi: 3.0.0
info:
  title: Financbase Admin Dashboard API
  version: 2.0.0
  description: >-
    A comprehensive financial management platform API with advanced automation,
    integrations, and analytics capabilities.
  contact:
    name: Financbase Support
    email: support@financbase.com
  license:
    name: Proprietary
    url: https://financbase.com/license
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://app.financbase.com
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk session token obtained from getToken()
security:
  - BearerAuth: []
...
paths:
...
  /api/auth/validate-token:
    post:
      summary: Validate Clerk JWT token for WebSocket connections
      description: >-
        Validates a Clerk JWT token and returns user information. Used by
        Cloudflare Workers to authenticate WebSocket connections.
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  description: Clerk session token obtained from getToken()
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    example: true
                  userId:
                    type: string
                    example: user_123
                  email:
                    type: string
                    example: user@example.com
        '400':
          description: Token is required
        '401':
          description: Token is invalid, expired, or user account is inactive
    get:
      summary: Health check for token validation endpoint
      description: Returns the status of the token validation endpoint
      tags:
        - Authentication
      responses:
        '200':
          description: Endpoint is active
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  message:
                    type: string
                    example: Token validation endpoint is active
...
  /api/search:
    get:
      summary: Search across the platform
      description: >-
        Performs a unified search across invoices, transactions, clients, pages,
        and reports
      tags:
        - Analytics
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
          description: Search query string
        - in: query
          name: type
          schema:
            type: string
            enum:
              - page
              - transaction
              - invoice
              - client
              - report
              - setting
          description: Filter results by type
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
          description: Maximum number of results to return
      responses:
        '200':
          description: Search results retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    title:
                      type: string
                    description:
                      type: string
                    type:
                      type: string
                      enum:
                        - page
                        - transaction
                        - invoice
                        - client
                        - report
                        - setting
                    href:
                      type: string
                    icon:
                      type: string
                    metadata:
                      type: object
        '401':
          description: Unauthorized - Authentication required
        '500':
          description: Internal server error
...